Fix reading images before checking for RGBA.

+++ mozillians/users/models.py
@@ -15,8 +15,9 @@ from django.utils.http import urlquote
 from django.template.loader import get_template
 
 from product_details import product_details
+from PIL import Image
 from pytz import common_timezones
-from sorl.thumbnail import ImageField, default as sorl_default, get_thumbnail
+from sorl.thumbnail import ImageField, get_thumbnail
 from django.utils.translation import ugettext as _, ugettext_lazy as _lazy
 
 from mozillians.common import utils
@@ -520,9 +521,9 @@ class UserProfile(UserProfilePrivacyModel):
         if 'crop' not in kwargs:
             kwargs['crop'] = 'center'
 
-        if self.photo:
+        if self.photo and default_storage.exists(self.photo):
             # Workaround for legacy images in RGBA model
-            image_obj = sorl_default.engine.get_image(self.photo)
+            image_obj = Image.open(self.photo)
             if image_obj.mode == 'RGBA':
                 new_fh = default_storage.open(self.photo.name, 'w')
                 converted_image_obj = image_obj.convert('RGB')
